requirements.txt
annotated-types==0.7.0
anyio==4.9.0
fastapi==0.116.1
greenlet==3.2.3
idna==3.10
passlib==1.7.4
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
